popup deleteSave

+++ EazySave-Master/View/RunSave.xaml.cs
@@ -80,8 +80,20 @@ namespace EazySave_Master.View
             // get number of save
             if (sender is Button button && button.DataContext is Save selectedSave)
             {
-                int selectedNumber = selectedSave.number;
-                m.mv.deleteSave(selectedNumber);
+
+                ResourceManager rm = m.mv.resourceManager;
+
+                MessageBoxResult result = MessageBox.Show(rm.GetString("PopupConfirmDelete"), "Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+                if (result == MessageBoxResult.Yes)
+                {
+
+                    int selectedNumber = selectedSave.number;
+                    m.mv.deleteSave(selectedNumber);
+
+                    button.Click -= DeleteButton_Click;
+
+                }
             }
         }
